Separate out core worker client interface from core worker client
