+++ src/ray/core_worker/BUILD.bazel
@@ -44,6 +44,7 @@ ray_cc_library(
         "//src/ray/pubsub:publisher",
         "//src/ray/pubsub:subscriber",
         "//src/ray/rpc:core_worker_client",
+        "//src/ray/rpc:core_worker_client_pool",
         "//src/ray/rpc:metrics_agent_client",
         "//src/ray/rpc:raylet_client_lib",
         "//src/ray/stats:stats_lib",
@@ -203,8 +204,7 @@ ray_cc_library(
         "//src/ray/protobuf:common_cc_proto",
         "//src/ray/pubsub:publisher_interface",
         "//src/ray/pubsub:subscriber_interface",
-        "//src/ray/rpc:core_worker_client",
-        "//src/ray/rpc:grpc_server",
+        "//src/ray/rpc:utils",
         "//src/ray/util:logging",
         "//src/ray/util:network_util",
         "@com_google_absl//absl/base:core_headers",
@@ -261,6 +261,7 @@ ray_cc_library(
         "//src/ray/common:ray_config",
         "//src/ray/common:status",
         "//src/ray/ipc:raylet_ipc_client_interface",
+        "//src/ray/rpc:utils",
         "@com_google_absl//absl/container:flat_hash_map",
         "@com_google_absl//absl/container:flat_hash_set",
         "@com_google_absl//absl/synchronization",
@@ -294,6 +295,7 @@ ray_cc_library(
         "//src/ray/common:protobuf_utils",
         "//src/ray/protobuf:common_cc_proto",
         "//src/ray/protobuf:core_worker_cc_proto",
+        "//src/ray/rpc:core_worker_client_interface",
         "//src/ray/stats:stats_metric",
         "//src/ray/util:counter_map",
         "//src/ray/util:exponential_backoff",
@@ -328,9 +330,8 @@ ray_cc_library(
     hdrs = ["future_resolver.h"],
     deps = [
         ":memory_store",
-        "//src/ray/common:grpc_util",
         "//src/ray/common:id",
-        "//src/ray/rpc:core_worker_client",
+        "//src/ray/rpc:core_worker_client_pool",
     ],
 )
