+++ src/ray/core_worker/task_execution/tests/task_receiver_test.cc
@@ -22,7 +22,7 @@
 #include "ray/common/asio/instrumented_io_context.h"
 #include "ray/common/task/task_spec.h"
 #include "ray/common/test_utils.h"
-#include "ray/rpc/worker/core_worker_client.h"
+#include "ray/rpc/worker/core_worker_client_interface.h"
 #include "ray/util/time.h"
 
 namespace ray {
