+++ src/ray/core_worker/task_submission/normal_task_submitter.h
@@ -31,7 +31,6 @@
 #include "ray/core_worker/task_submission/dependency_resolver.h"
 #include "ray/rpc/raylet/raylet_client_interface.h"
 #include "ray/rpc/raylet/raylet_client_pool.h"
-#include "ray/rpc/worker/core_worker_client.h"
 #include "ray/rpc/worker/core_worker_client_pool.h"
 
 namespace ray {
