+++ src/ray/core_worker/task_submission/tests/BUILD.bazel
@@ -53,7 +53,7 @@ ray_cc_test(
         "//src/ray/core_worker:fake_actor_creator",
         "//src/ray/core_worker:reference_count",
         "//src/ray/core_worker:task_manager",
-        "//src/ray/rpc:core_worker_client",
+        "//src/ray/rpc:core_worker_client_interface",
         "@com_google_googletest//:gtest",
         "@com_google_googletest//:gtest_main",
     ],
@@ -71,7 +71,8 @@ ray_cc_test(
         "//src/ray/core_worker:fake_actor_creator",
         "//src/ray/core_worker:memory_store",
         "//src/ray/core_worker/task_submission:normal_task_submitter",
-        "//src/ray/rpc:core_worker_client",
+        "//src/ray/rpc:core_worker_client_interface",
+        "//src/ray/rpc:core_worker_client_pool",
         "//src/ray/rpc:fake_raylet_client",
         "//src/ray/rpc:raylet_client_interface",
         "@com_google_googletest//:gtest",
