+++ src/ray/core_worker/task_submission/tests/actor_task_submitter_test.cc
@@ -23,7 +23,7 @@
 #include "mock/ray/core_worker/task_manager_interface.h"
 #include "ray/common/test_utils.h"
 #include "ray/core_worker/fake_actor_creator.h"
-#include "ray/rpc/worker/core_worker_client.h"
+#include "ray/rpc/worker/core_worker_client_interface.h"
 
 namespace ray::core {
