+++ src/ray/gcs/BUILD.bazel
@@ -256,7 +256,7 @@ ray_cc_library(
         "//src/ray/observability:ray_driver_job_execution_event",
         "//src/ray/observability:ray_event_recorder_interface",
         "//src/ray/pubsub:gcs_publisher",
-        "//src/ray/rpc:core_worker_client",
+        "//src/ray/rpc:core_worker_client_pool",
         "//src/ray/stats:stats_metric",
         "//src/ray/util:event",
         "//src/ray/util:thread_checker",
@@ -392,7 +392,7 @@ ray_cc_library(
         "//src/ray/common:id",
         "//src/ray/common:ray_config",
         "//src/ray/raylet/scheduling:cluster_lease_manager",
-        "//src/ray/rpc:core_worker_client",
+        "//src/ray/rpc:core_worker_client_pool",
         "//src/ray/rpc:raylet_client_interface",
         "//src/ray/rpc:raylet_client_pool",
         "//src/ray/util:logging",
@@ -426,7 +426,8 @@ ray_cc_library(
         "//src/ray/common:task_common",
         "//src/ray/protobuf:gcs_service_cc_proto",
         "//src/ray/pubsub:gcs_publisher",
-        "//src/ray/rpc:core_worker_client",
+        "//src/ray/rpc:core_worker_client_interface",
+        "//src/ray/rpc:core_worker_client_pool",
         "//src/ray/stats:stats_lib",
         "//src/ray/util:counter_map",
         "//src/ray/util:logging",
@@ -514,6 +515,7 @@ ray_cc_library(
         "//src/ray/pubsub:publisher",
         "//src/ray/raylet/scheduling:scheduler",
         "//src/ray/rpc:core_worker_client",
+        "//src/ray/rpc:core_worker_client_pool",
         "//src/ray/rpc:grpc_server",
         "//src/ray/rpc:metrics_agent_client",
         "//src/ray/rpc:raylet_client_lib",
