+++ src/ray/gcs/gcs_actor_manager.h
@@ -35,7 +35,7 @@
 #include "ray/gcs/grpc_service_interfaces.h"
 #include "ray/gcs/usage_stats_client.h"
 #include "ray/pubsub/gcs_publisher.h"
-#include "ray/rpc/worker/core_worker_client.h"
+#include "ray/rpc/worker/core_worker_client_interface.h"
 #include "ray/rpc/worker/core_worker_client_pool.h"
 #include "ray/util/counter_map.h"
 #include "ray/util/thread_checker.h"
