+++ src/ray/gcs/tests/BUILD.bazel
@@ -131,6 +131,7 @@ ray_cc_test(
         "//src/ray/gcs:gcs_kv_manager",
         "//src/ray/gcs/store_client:in_memory_store_client",
         "//src/ray/observability:fake_ray_event_recorder",
+        "//src/ray/rpc:core_worker_client_pool",
         "@com_google_googletest//:gtest_main",
     ],
 )
@@ -233,6 +234,7 @@ ray_cc_test(
         "//src/ray/common:test_utils",
         "//src/ray/gcs:gcs_actor",
         "//src/ray/gcs:gcs_actor_scheduler",
+        "//src/ray/rpc:core_worker_client_pool",
         "//src/ray/util:counter_map",
         "@com_google_googletest//:gtest_main",
     ],
@@ -423,7 +425,8 @@ ray_cc_test(
         "//src/ray/gcs:gcs_function_manager",
         "//src/ray/gcs/store_client:in_memory_store_client",
         "//src/ray/pubsub:publisher",
-        "//src/ray/rpc:core_worker_client",
+        "//src/ray/rpc:core_worker_client_interface",
+        "//src/ray/rpc:core_worker_client_pool",
         "//src/ray/util:event",
         "@com_google_googletest//:gtest_main",
     ],
