+++ src/ray/gcs/tests/gcs_actor_scheduler_mock_test.cc
@@ -26,6 +26,7 @@
 #include "ray/gcs/gcs_actor.h"
 #include "ray/gcs/gcs_actor_scheduler.h"
 #include "ray/observability/fake_ray_event_recorder.h"
+#include "ray/rpc/worker/core_worker_client_pool.h"
 #include "ray/util/counter_map.h"
 
 using namespace ::testing;  // NOLINT
