+++ src/ray/object_manager/ownership_object_directory.cc
@@ -18,7 +18,6 @@
 #include <string>
 #include <unordered_set>
 #include <utility>
-#include <vector>
 
 #include "ray/stats/metric_defs.h"
