+++ src/ray/object_manager/ownership_object_directory.h
@@ -25,8 +25,7 @@
 #include "ray/common/status.h"
 #include "ray/gcs_client/gcs_client.h"
 #include "ray/object_manager/object_directory.h"
-#include "ray/pubsub/subscriber.h"
-#include "ray/rpc/worker/core_worker_client.h"
+#include "ray/pubsub/subscriber_interface.h"
 #include "ray/rpc/worker/core_worker_client_pool.h"
 #include "ray/stats/metric.h"
