+++ src/ray/raylet/BUILD.bazel
@@ -89,12 +89,12 @@ ray_cc_library(
         "//src/ray/common:lease",
         "//src/ray/flatbuffers:node_manager_generated",
         "//src/ray/ipc:client_connection",
+        "//src/ray/protobuf:core_worker_cc_grpc",
+        "//src/ray/protobuf:core_worker_cc_proto",
         "//src/ray/raylet/scheduling:cluster_resource_scheduler",
         "//src/ray/rpc:core_worker_client",
         "//src/ray/util:process",
-        "@com_google_absl//absl/memory",
         "@com_google_absl//absl/time",
-        "@com_google_googletest//:gtest_prod",
     ],
 )
 
@@ -163,12 +163,10 @@ ray_cc_library(
         ":worker_pool",
         "//src/ray/common:id",
         "//src/ray/common:ray_object",
-        "//src/ray/gcs_client",
         "//src/ray/object_manager:object_directory",
-        "//src/ray/object_manager:object_manager_common",
         "//src/ray/protobuf:node_manager_cc_proto",
         "//src/ray/pubsub:subscriber_interface",
-        "//src/ray/rpc:core_worker_client",
+        "//src/ray/rpc:core_worker_client_pool",
         "//src/ray/util:time",
     ],
 )
@@ -290,6 +288,8 @@ ray_cc_binary(
         "//src/ray/gcs_client",
         "//src/ray/object_manager:ownership_object_directory",
         "//src/ray/raylet/scheduling:cluster_lease_manager",
+        "//src/ray/rpc:core_worker_client",
+        "//src/ray/rpc:core_worker_client_pool",
         "//src/ray/rpc:metrics_agent_client",
         "//src/ray/rpc:raylet_client_lib",
         "//src/ray/rpc:raylet_client_pool",
