+++ src/ray/raylet/main.cc
@@ -40,6 +40,8 @@
 #include "ray/raylet/raylet.h"
 #include "ray/rpc/object_manager/object_manager_client.h"
 #include "ray/rpc/raylet/raylet_client.h"
+#include "ray/rpc/worker/core_worker_client.h"
+#include "ray/rpc/worker/core_worker_client_pool.h"
 #include "ray/stats/stats.h"
 #include "ray/stats/tag_defs.h"
 #include "ray/util/cmd_line_utils.h"
