+++ src/ray/raylet/tests/BUILD.bazel
@@ -56,7 +56,8 @@ ray_cc_test(
         "//src/ray/pubsub:subscriber",
         "//src/ray/raylet:local_object_manager",
         "//src/ray/raylet:worker_pool",
-        "//src/ray/rpc:core_worker_client",
+        "//src/ray/rpc:core_worker_client_interface",
+        "//src/ray/rpc:core_worker_client_pool",
         "//src/ray/rpc:grpc_client",
         "@com_google_googletest//:gtest_main",
     ],
@@ -191,6 +192,7 @@ ray_cc_test(
         "//src/ray/raylet:local_object_manager_interface",
         "//src/ray/raylet:node_manager",
         "//src/ray/raylet/scheduling:cluster_lease_manager",
+        "//src/ray/rpc:core_worker_client_pool",
         "//src/ray/rpc:fake_raylet_client",
         "//src/ray/util:macros",
         "@com_google_googletest//:gtest_main",
