+++ src/ray/rpc/BUILD.bazel
@@ -194,27 +194,63 @@ ray_cc_library(
 )
 
 ray_cc_library(
+    name = "core_worker_client_interface",
+    hdrs = [
+        "worker/core_worker_client_interface.h",
+    ],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//src/ray/protobuf:core_worker_cc_proto",
+        "//src/ray/protobuf:pubsub_cc_proto",
+        "//src/ray/pubsub:subscriber_interface",
+    ],
+)
+
+ray_cc_library(
     name = "core_worker_client",
     srcs = [
         "worker/core_worker_client.cc",
-        "worker/core_worker_client_pool.cc",
     ],
     hdrs = [
         "worker/core_worker_client.h",
+    ],
+    deps = [
+        ":core_worker_client_interface",
+        ":retryable_grpc_client",
+        "//src/ray/protobuf:core_worker_cc_grpc",
+        "//src/ray/protobuf:core_worker_cc_proto",
+        "//src/ray/util:logging",
+        "@com_google_absl//absl/synchronization",
+    ],
+)
+
+ray_cc_library(
+    name = "core_worker_client_pool",
+    srcs = [
+        "worker/core_worker_client_pool.cc",
+    ],
+    hdrs = [
         "worker/core_worker_client_pool.h",
     ],
     deps = [
-        ":raylet_client_interface",
+        ":core_worker_client_interface",
         ":raylet_client_pool",
         "//src/ray/common:id",
         "//src/ray/common:status",
         "//src/ray/gcs_client",
-        "//src/ray/protobuf:core_worker_cc_grpc",
-        "//src/ray/pubsub:subscriber",
         "//src/ray/util:logging",
         "//src/ray/util:network_util",
-        "@com_github_grpc_grpc//:grpc++",
         "@com_google_absl//absl/synchronization",
+    ],
+)
+
+ray_cc_library(
+    name = "utils",
+    hdrs = ["utils.h"],
+    visibility = ["//visibility:public"],
+    deps = [
+        "//src/ray/protobuf:common_cc_proto",
+        "@com_google_protobuf//:protobuf",
     ],
 )
 
@@ -225,7 +261,7 @@ ray_cc_library(
     ],
     visibility = ["//visibility:public"],
     deps = [
-        ":core_worker_client",
+        ":core_worker_client_interface",
         "@com_google_absl//absl/synchronization",
     ],
 )
