+++ src/ray/rpc/tests/core_worker_client_pool_test.cc
@@ -23,7 +23,7 @@
 
 #include "gmock/gmock.h"
 #include "mock/ray/raylet_client/raylet_client.h"
-#include "ray/rpc/worker/core_worker_client.h"
+#include "ray/rpc/worker/core_worker_client_interface.h"
 
 namespace ray {
 namespace rpc {
