+++ src/ray/rpc/worker/core_worker_client_pool.h
@@ -16,7 +16,6 @@
 
 #include <list>
 #include <memory>
-#include <string>
 #include <utility>
 
 #include "absl/base/thread_annotations.h"
@@ -26,11 +25,12 @@
 #include "ray/gcs_client/gcs_client.h"
 #include "ray/rpc/raylet/raylet_client_interface.h"
 #include "ray/rpc/raylet/raylet_client_pool.h"
-#include "ray/rpc/worker/core_worker_client.h"
+#include "ray/rpc/worker/core_worker_client_interface.h"
 
 namespace ray {
 namespace rpc {
-
+using CoreWorkerClientFactoryFn =
+    std::function<std::shared_ptr<CoreWorkerClientInterface>(const rpc::Address &)>;
 class CoreWorkerClientPool {
  public:
   CoreWorkerClientPool() = delete;
